+++ NEWS.md
@@ -1,5 +1,8 @@
 # pkgload (development version)
 
+* Fix handling of active bindings inside a package during unloading (#255, @klmr).
+
+
 # pkgload 1.3.3
 
 * pkgload now depends unconditionally on pkgbuild (#249).